perf(core) loop over the plugins on startup only once
### Summary

When there is a database with a huge amounts of plugins the startup can be
a bit slow as we loop through plugins three times:

1. on init to check if there is plugin configured that is not installed on node
2. on init to build initial plugin iterator
3. on init_worker to pre-warm cache with plugin entities

This commit combines them together so that we only loop through plugins once.

This is still WIP as I need to figure out how to properly combine init/init_worler,
the remaining problem is that `cache` is not available on `init` while it is
available with `init_worker`.